IGNITE-16113 Support for BINARY/VARBINARY data types

+++ modules/calcite/src/test/java/org/apache/ignite/internal/processors/query/calcite/exec/exp/IgniteSqlFunctionsTest.java
@@ -28,7 +28,7 @@ public class IgniteSqlFunctionsTest {
     /** */
     @Test
     public void testBigDecimalToString() {
-        Assert.assertNull(IgniteSqlFunctions.toString(null));
+        Assert.assertNull(IgniteSqlFunctions.toString((BigDecimal)null));
 
         Assert.assertEquals(
             "10",
